+++ src/app/Components/ComponentsSamples/Navigation/FloatingActionButtonSamplesViewModel.cs
@@ -1,0 +1,14 @@
+using System.Windows.Input;
+
+namespace Components.ComponentsSamples.Navigation;
+
+public class FloatingActionButtonSamplesViewModel
+{
+    public FloatingActionButtonSamplesViewModel()
+    {
+        NavigateToFloatingNavigationButtonSamples = new Command(() =>
+            Shell.Current.Navigation.PushAsync(new FloatingNavigationButtonSamples()));
+    }
+    
+    public ICommand NavigateToFloatingNavigationButtonSamples { get; }
+}