+++ src/app/Components/Platforms/Android/MainActivity.cs
@@ -11,4 +11,5 @@ public class MainActivity : MauiAppCompatActivity
     public MainActivity()
     {
     }
+    
 }