+++ src/app/Playground/AppShell.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace Playground;
+
+public partial class AppShell : Shell
+{
+    public AppShell()
+    {
+        InitializeComponent();
+    }
+}