+++ src/app/Playground/VetleSamples/VetlePage.xaml.cs
@@ -1,0 +1,13 @@
+using System.Windows.Input;
+
+namespace Playground.VetleSamples;
+
+public partial class VetlePage
+{
+    public VetlePage()
+    {
+        InitializeComponent();
+    }
+    
+
+}