+++ src/library/DIPS.Mobile.UI/Components/Navigation/FloatingNavigationButton/IFloatingNavigationButtonConfigurator.cs
@@ -1,0 +1,9 @@
+using System.Windows.Input;
+
+namespace DIPS.Mobile.UI.Components.Navigation.FloatingNavigationButton;
+
+public interface IFloatingNavigationButtonConfigurator
+{
+    void AddNavigationButton(string identifier, string title, IconName iconName, ICommand command);
+    void AddPageThatHidesButton(Type page);
+}