+++ src/library/DIPS.Mobile.UI/Components/Navigation/FloatingNavigationButton/dotnet/FloatingNavigationButtonService.cs
@@ -1,0 +1,9 @@
+// ReSharper disable once CheckNamespace
+namespace DIPS.Mobile.UI.Components.Navigation.FloatingNavigationButton;
+
+public partial class FloatingNavigationButtonService
+{
+    private static partial void AttachToRootWindow(FloatingNavigationButton fab)
+    {
+    }
+}