+++ src/library/DIPS.Mobile.UI/Effects/Touch/Touch.Properties.cs
@@ -1,22 +1,22 @@
 using System.Windows.Input;
 
-namespace DIPS.Mobile.UI.Effects.DUITouchEffect;
+namespace DIPS.Mobile.UI.Effects.Touch;
 
-public partial class DUITouchEffect
+public partial class Touch
 {
     public static readonly BindableProperty CommandProperty = BindableProperty.CreateAttached("Command",
         typeof(ICommand),
-        typeof(DUITouchEffect),
+        typeof(Touch),
         null,
         propertyChanged: OnCommandChanged);
     
     public static readonly BindableProperty CommandParameterProperty = BindableProperty.CreateAttached("Command",
         typeof(object),
-        typeof(DUITouchEffect),
+        typeof(Touch),
         null);
     
     public static readonly BindableProperty AccessibilityContentDescriptionProperty = BindableProperty.CreateAttached("AccessibilityContentDescription",
         typeof(string),
-        typeof(DUITouchEffect),
+        typeof(Touch),
         null);
 }