+++ src/library/DIPS.Mobile.UI/Effects/Touch/TouchPlatformEffect.cs
@@ -1,8 +1,8 @@
 using Microsoft.Maui.Controls.Platform;
 
-namespace DIPS.Mobile.UI.Effects.DUITouchEffect;
+namespace DIPS.Mobile.UI.Effects.Touch;
 
-public partial class DUITouchPlatformEffect : PlatformEffect
+public partial class TouchPlatformEffect : PlatformEffect
 {
     protected override partial void OnAttached();
     protected override partial void OnDetached();